Improvents discussed in paper
As Zemlin and I discussed yesterday going over the report I have
corrected the formula to use the half way point of a color's cumulative
percent rather than the last point of the color's cumulative percent. In
a side by side comparison the resulting photo was slightly darker and
probably has a slightly more equalized histogram.